[Garima] refactor drop down component for modal being opened
