+++ .eslintrc.js
@@ -54,5 +54,4 @@ module.exports = {
     "jest/valid-title": "error",
     "jest/valid-expect": "error", "jest/valid-expect-in-promise": "error"
   }
-
 };
