+++ assets/theme/colors.js
@@ -1,4 +1,8 @@
 const Dim_Black ="#424242";
+const Black = "#3E4059";
+const White = "#ffff";
 export const colors = {
-  Dim_Black
+  Dim_Black,
+  Black,
+  White
 };
